Fix binary to 'jive5ab', StartJ5 not installed
It's getting a bit messy, there is overlap between jive-python-utils
package (still is now) but the StartJ5 script is mostly there for local
tuning options to persist so should not be (re)installed/overwritten.

Also the binary name contained explicit version info, which makes
external scripts (e.g. systemd bootup scripts) basically impossible to
figure out which binary to start. The old 'solution' (i.e. maintaining
symlinks to the actual binary) is quite messy - when choosing for a
package management system, one expects the package management system to
keep track of the version.